quant-statarb-kalman/pyproject.toml
[project]
name = "quant-statarb-kalman"
version = "0.1.0"
description = "Sector-neutral stat-arb with Kalman dynamic hedge and walk-forward backtest"
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "scipy>=1.10",
  "statsmodels>=0.14",
  "yfinance>=0.2",
  "requests>=2.31",
  "pyarrow>=14.0",
]

[project.scripts]
statarb = "statarb.cli:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["statarb*", "execution*"]

[tool.pytest.ini_options]
addopts = "-q"
